requirements.txt
anthropic>=0.40.0
watchdog>=4.0.0
Pillow>=10.0.0
